generalized api for scraping relevant sets
