+++ dataset-scraping/output.csv
@@ -1,0 +1,70 @@
+"name","description","category","url"
+"APD Crime","From https://github.com/codeforatlanta/apd-crime-data/blob/master/APD_crime.csv","PUBLIC SAFETY",
+"Poverty Estimates","","",
+"Atlanta City Planning: Land Quadrants","","",
+"SNAP Data: Estimates + Margin of Error","RECEIPT OF FOOD STAMPS/SNAP IN THE PAST 12 MONTHS BY POVERTY STATUS IN THE PAST 12 MONTHS FOR HOUSEHOLDS\nUniverse: Households  more information\n2008-2012 American Community Survey 5-Year Estimates\n\nSource - US Census (Census.gov)","",
+"Map of Georgia Polling Places","Current locations of all polling places in the state","CIVIC ENGAGEMENT",
+"Poverty Estimates by County","","",
+"Kroger Food Markets (Locations)","","",
+"Georgia Polling Places","Current locations of all polling places in the state","CIVIC ENGAGEMENT",
+"School Ranking Output","","EDUCATION",
+"Demographics for Georgia Schools Pre-K through 12th grade for 2013","","EDUCATION",
+"Schools & Data","Schools of GA, their location, FRL and CRCT for 2011 2014","EDUCATION",
+"SchoolsAndData","Schools of GA, their location, FRL and CRCT for 2011 2014","EDUCATION",
+"Walmart Stores (Locations near ATL)","","",
+"SNAP Data by Census Tract 2008-2012","RECEIPT OF FOOD STAMPS/SNAP IN THE PAST 12 MONTHS BY POVERTY STATUS IN THE PAST 12 MONTHS FOR HOUSEHOLDS\nUniverse: Households  more information\n2008-2012 American Community Survey 5-Year Estimates\n\nSource - US Census (Census.gov)","",
+"Georgia Schools with location","","EDUCATION",
+"Atlanta DPW Work Order Report OCT 2013 - OCT 2014","Work orders from the Atlanta Department of Public Works","CITY PROCESS  IMPROVEMENTS",
+"Fulton County Food Service Inspection Scores (location column separated)","Health inspections of food services in Fulton County 2014-2015","HEALTH",
+"Walmart Stores in the United States & Canada (Locations)","","",
+"Fulton County Food Service Inspection Scores 2014-2015","Health inspections of food services in Fulton County 2014-2015","HEALTH",
+"Publix Stores (Locations)","","",
+"Rite Aid Stores (Locations near ATL)","","",
+"Farmers Markets","","",
+"Rite Aid Stores (Locations)","","",
+"Georgia zipcodes with centroids","","",
+"Kroger Food Markets (Locations near ATL)","","",
+"Publix Stores (Locations near ATL)","","",
+"Poverty Rates for Counties in Georgia (Chart)","","",
+"Acfb Inventory","A proof of concept dataset from the TEST HARVEST acfb server.","OPEN DATA SHARING PLATFORM",
+"Socrata See Click Fix Data Huntsville","","311",
+"Community Resources: Count by Type","","",
+"Atlanta Neighborhood Planning Units (NPUs)","","NEIGHBORHOODS",
+"Atlanta Community Resources","","",
+"Map of Atlanta Community Resources","","",
+"Poverty Estimates by County (GA)","","",
+"TEST DATA: SNAP Discounts by Store, User, Date","","",
+"SCI Neighborhood Survey Totals","This dataset is a neighborhood by neighborhood accounting of parcels, vacant parcels, structure state, and other data points related to housing in Atlanta.","HOUSING",
+"Fulton County Tax Landlots","","",
+"Atlanta Neighborhoods","","",
+"Poverty Rates by State (Map)","","",
+"Brigade - Code for Atlanta - Info","","OPEN DATA SHARING PLATFORM",
+"Greater Atlanta Food Deserts","","",
+"Major Roads (Atlanta, GA)","This layer was developed by the Research & Analytics Division of the Atlanta Regional Commission and is a subset of the Georgia Department of Transportation's (GDOT) DLGF street centerline database. The features included in this Layer consist of all State highways in the 29-county area, as well as a number of additional roads that were identified by ARC's Transportation Planning Division (TPD) as major roads. Please note, this Layer is intended for relatively small scale mapping and labeling, and should be used in conjunction with the Expressways Layer. For more information on the DLGF streetbase, refer to the metadata for ARstreets and Streets.","",
+"Total Discounts by SNAP User","","",
+"Pat - Test","","",
+"Total Sales and Total Discounts by Store","","",
+"Atlanta City Council Districts (2014)","","",
+"Voting Wait Times","crowdsourced wait times at voting precincts","OPEN DATA SHARING PLATFORM",
+"Georgia Early Voting Locations for 2015/06/16 Special Election","","CIVIC ENGAGEMENT",
+"Atlanta City Planning: Zoning Base","","",
+"Georgia general election, 2014/11/04","","CIVIC ENGAGEMENT",
+"Poverty Estimates by State","","",
+"Atlanta Neighborhoods + Grocery Stores","","",
+"Atlanta City Limits","","",
+"Total Weight, Sales, and Discounts by Produce Code","","",
+"Address GIS data for Fulton County, GA and the City of Atlanta","Address points are intended as a comprehensive collection of all addresses in Fulton County and the City of Atlanta represented as point locations. The locations generally represent the drveway access to properties, thus the commonly used name 'Driveway Access Points', or 'Daccess' for short. Address attributes include the individual components of the address, the status (active or inactive), and the date the address was entered into the database. The primary source of address points are development plats approved through Fulton County and the City of Atlanta. Addresses are also updated as a result of address changes processed by the County and it's municipalities. The address points are maintained primarily by Fulton County GIS, with contributions from the GIS Division of the City of Atlanta Department of Planning and Community Development.","HOUSING",
+"Brigade - Code for Atlanta - Topics","","OPEN DATA SHARING PLATFORM",
+"CVS","Locations of CVS stores in Atlanta","",
+"Georgia Ports","US Army Corps of Engineers Data - Filtered by state of Georgia\nNational Transportation Atlas Database 2012","",
+"Fulton County Trans Centerline","","",
+"20150714-locations","","CIVIC ENGAGEMENT",
+"Atlanta City Planning: Zoning Overlay","","",
+"Atlanta Bicycle Facilities","","",
+"U.S. Army Corps of Engineers Ports","","TRANSPORTATION",
+"Atlanta City Planning: Landmark Sites","","",
+"Georgia Voting Precincts + Polling Places 2014","Fetch latest precinct data every year from the Census Bureau","PLANNING",
+"Georgia Voting Precincts 2014","Fetch latest precinct data every year from the Census Bureau","PLANNING",
+"Reported Crimes - Cleaned","","PUBLIC SAFETY",
+"Map View","","PUBLIC SAFETY",
+"Atlanta Map","","PUBLIC SAFETY",
